both patch template and user/test database
