+++ Makefile
@@ -98,7 +98,7 @@ assets:
 	mkdir -p kalite/database/templates/
 	cp kalite/database/data.sqlite kalite/database/templates/
 	bin/kalite manage retrievecontentpack download en --minimal --foreground --template
-	bin/kalite manage patch_content_db_template
+	bin/kalite manage patch_content_db --template
 
 release: dist man
 	ls -l dist
