+++ kalite/testing/testrunner.py
@@ -144,6 +144,7 @@ class KALiteTestRunner(DjangoTestSuiteRunner):
 
         if not database_exists():
             call_command("retrievecontentpack", "download", "en", minimal=True, foreground=True)
+            call_command("patch_content_db")
             logging.info("Successfully setup content database")
         else:
             logging.info("Content database already exists")
